\name{fsurf.var.ls}
\alias{fsurf.var.ls}
\title{Generate variable names to match fsurf summary}
\description{}
\usage{
  fsurf.var.ls(rois = c("peg","all"), hemi = c("t", "r", "l"), vars = c("sa", "g", "ta", "mc"))
}
\arguments{
  \item{rois}{}
  \item{hemi}{}
  \item{vars}{}
}
\value{
  \item{var.names}{}
}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}