\name{fsurfR-package}
\alias{fsurfR-package}
\alias{fsurfR}
\docType{package}
\title{
\packageTitle{fsurfR}
}
\description{
\packageDescription{fsurfR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fsurfR}
\packageIndices{fsurfR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fsurfR}

Maintainer: \packageMaintainer{fsurfR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
