\name{init.curv}
\alias{init.curv}
\title{Initialize empty CURV file}
\description{}
\usage{
  init.curv(file.name, surf.file)
}
\arguments{
  \item{curv.file}{ Full file path to desired CURV file. While not overwrite if file already exists.}
  \item{surf.file}{ SURF file corresponding to the CURV file you wish to initialize. This file sets the number of verteices, etc.}
}
\value{A file will be created in the desired directory.}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}