\name{load.vertex}
\alias{load.vertex}
\title{Load vertices for group analysis}
\description{Load vertex values for all subjects, assuming they have been resample to a common space.}
\usage{
  load.vertex(data.dir, which.sjx = "all", hemi = c("lh", "rh"), var.name = c("curv", "sulc", "thickness", "volume", "w-g.pct", "jacobian_white"), which.vertex)
}
\arguments{
  \item{data.dir}{}
  \item{which.sjx}{}
  \item{hemi}{}
  \item{var.name}{}
  \item{which.vertex}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}