\name{mgh.hdr}
\alias{mgh.hdr}
\title{Retreive MGH File Header Information}
\description{}
\usage{
  mgh.hdr(mgh.file, field = "all")
}
\arguments{
  \item{mgh.file}{}
  \item{field}{"all", c("version", "width", "height", "depth", "nframes", "type", "dof", "goodRASFlag", "spacing", "x_ras", "y_ras", "z_ras", "c_ras")}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}