\name{read.curv.vertex}
\alias{read.curv.vertex}
\title{Read values from CURV file}
\description{}
\usage{
  read.curv.vertex(curv.file, n="all", surf.file=NULL, coords=NULL)
}
\arguments{
  \item{curv.file}{}
  \item{n}{}
  \item{surf.file}{}
  \item{coords}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}