\name{read.fsurf.stats}
\alias{read.fsurf.stats}
\title{Read Freesurfer Stats File (skipping header differences between versions)}
\description{}
\usage{
  read.fsurf.stats(stats.file)
}
\arguments{
  \item{stats.file}{}
}
\value{
}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}