\name{read.surf}
\alias{read.surf}
\title{Read values from SURF file}
\description{}
\usage{
  read.surf(surf.file, which.vertex, which.face)
}
\arguments{
  \item{surf.file}{}
  \item{which.vertex}{}
  \item{which.face}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}