\name{summary.fsurf}
\alias{summary.fsurf}
\title{summarize surf variables across subjects}
\description{}
\usage{
  summary.fsurf(data.dir, vars = "all", rois = "peg", sjx = "all", hemi = c("t", "l", "r"), save.csv=TRUE, save.dir = NULL, return.df = FALSE)
}
\arguments{
  \item{data.dir}{}
  \item{vars}{}
  \item{rois}{}
  \item{sjx}{}
  \item{hemi}{}
  \item{save.csv}{}
  \item{save.dir}{}
  \item{return.df}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}