\name{write.curv.vertex}
\alias{write.curv.vertex}
\title{Write values to CURV file}
\description{}
\usage{
  write.curv.vertex(curv.file, n, values, surf.file=NULL, coords=NULL)
}
\arguments{
  \item{curv.file}{}
  \item{n}{}
  \item{surf.file}{}
  \item{coords}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}