\name{write.mgh.voxel}
\alias{write.mgh.voxel}
\title{MGH voxelwise write}
\description{Write voxelwise data to MGH file}
\usage{
  write.mgh.voxel(mgh.file, coords, value)
}
\arguments{
  \item{mgh.file}{}
  \item{coords}{}
  \item{value}{}
}
\value{}
\author{
  Timothy R. Koscik <timkoscik+fsurfR@gmail.com>
}
\examples{}